sourcmap tools: replace SourceMapUploader with SymbolUploader
