+++ tools/sourcemap-tools/src/index.ts
@@ -1,4 +1,4 @@
 export * from './ContentAppender';
 export * from './DebugIdGenerator';
-export * from './SourceMapUploader';
 export * from './SourceProcessor';
+export * from './SymbolUploader';
